requirements.txt
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1